+++ india_compliance/public/js/india_compliance_account/pages/AccountPage.vue
@@ -31,10 +31,10 @@
           <ul class="links">
             <a @click.prevent="showUsage"><li>Review API Usage</li></a>
             <!-- <a href="#"><li>Check API Status</li></a> -->
+            <a @click.prevent="openInvoiceDialog"><li>Invoice History</li></a>
             <a href="https://discuss.erpnext.com/c/erpnext/india-compliance/65"><li>Community Forum</li></a>
             <a href="https://github.com/resilient-tech/india-compliance/issues/new"><li>Report a Bug</li></a>
             <a href="mailto:api-support@indiacompliance.app"><li>Email Support</li></a>
-            <a @click.prevent="openInvoiceDialog"><li>Invoice History</li></a>
             <a @click.prevent="logout"><li>Logout</li></a>
           </ul>
         </div>
@@ -47,7 +47,10 @@
 import PageTitle from "../components/PageTitle.vue";
 import Message from "../components/Message.vue";
 import PreLoader from "../components/PreLoader.vue";
-import { getReadableNumber, get_invoice_history_dialog } from "../utils";
+import { getReadableNumber } from "../utils";
+import { get_invoice_history, send_invoice_email } from '../services/AccountService';
+import "../../../../templates/invoice_history_table.html";
+
 
 export default {
   components: {
@@ -80,8 +83,86 @@ data() {
         }
       );
     },
-    openInvoiceDialog(){
-      get_invoice_history_dialog(this.subscriptionDetails.email).show();
+    openInvoiceDialog() {
+      const invoiceHistoryDialog = new frappe.ui.Dialog({
+        title: __("Invoice History"),
+        fields: [
+          {
+            fieldname: "from_date",
+            label: __("From Date"),
+            fieldtype: "Date",
+            reqd: 1,
+            default: frappe.datetime.add_months(frappe.datetime.get_today(), -12),
+            onchange: () => {
+              console.log('onchange from_date');
+            },
+          },
+          {
+            fieldname: "email",
+            label: __("Email"),
+            fieldtype: "Data",
+            description: __("Invoice will be sent to this email address"),
+            options: "Email",
+            default: this.subscriptionDetails.email,
+          },
+          {
+            fieldtype: "Column Break"
+          },
+          {
+            fieldname: "to_date",
+            label: __("To Date"),
+            fieldtype: "Date",
+            reqd: 1,
+            default: frappe.datetime.get_today(),
+            onchange: () => {
+              console.log('onchange to_date');
+            },
+          },
+          {
+            label: __("Invoice History"),
+            fieldtype: "Section Break",
+          },
+          {
+            fieldname: "invoice_history",
+            label: __("Invoice History"),
+            fieldtype: "HTML",
+            hidden: 1,
+          }
+        ],
+        primary_action_label: __("Get Invoice History"),
+        primary_action: async (values) => {
+          const { from_date, to_date } = values;
+
+          if (from_date > to_date)
+            frappe.throw(__("From Date cannot be greater than To Date"));
+
+          const response = await get_invoice_history(from_date, to_date);
+          const data = response.message;
+
+          const invoiceHistoryTable = frappe.render_template("invoice_history_table", { data_array: data.length > 0 ? data : null });
+          invoiceHistoryDialog.fields_dict.invoice_history.html(invoiceHistoryTable);
+
+          invoiceHistoryDialog.fields_dict.invoice_history.df.hidden = 0;
+
+          invoiceHistoryDialog.fields_dict.invoice_history.$wrapper.ready(function () {
+            $('.get-invoice').click(async function () {
+              const invoice_name = $(this).data('invoice-name');
+              const email = invoiceHistoryDialog.get_value('email');
+
+              const response = await send_invoice_email(invoice_name, email);
+              if (response.success) {
+                frappe.msgprint({
+                  title: __("Success"),
+                  message: __("Invoice {0} sent successfully.", [invoice_name]),
+                  indicator: "green",
+                });
+              }
+            });
+          });
+          invoiceHistoryDialog.refresh();
+        },
+      });
+      invoiceHistoryDialog.show();
     }
   },
   computed: {
